added the strata extract logic into one file
